Let vmux_native build for iOS
The crate is 2,861 lines and exactly 199 of them knew what platform they
were on. surface.rs asked an NSView to order itself among siblings, take
first responder, carry a zPosition and override its appearance; the dom,
the shim, the routing, the measurement and the event plumbing never
mentioned AppKit at all. So the whole view half was gated to macOS on
account of four methods.

Those four move to surface/macos.rs and surface/ios.rs, the way
native_page.rs already splits. UIKit answers all four, and two of them
more simply than AppKit does: every UIView is layer-backed from birth, so
raise_above_layers has no setWantsLayer and no absent-layer path, and a
UIResponder is told to become first responder rather than asking a window
to hand it over.

The rest of the crate swaps target_os = "macos" for the `ui` alias, which
is what that alias was for. It needs a build.rs to emit it. Cargo cannot
read the alias -- dependency resolution runs before any build script --
so the manifest spells the same predicate out, which is the agreement
build_platform_cfg.rs documents.

Nothing calls the iOS half yet. vmux_browser is the only consumer and it
is macOS-only; this is what has to be true before the phone can have a
PageSurface at all.
