+++ crates/vmux_native/build.rs
@@ -1,0 +1,9 @@
+//! Emits the platform cfg aliases this crate splits on.
+
+#[path = "../build_platform_cfg.rs"]
+mod build_platform_cfg;
+
+fn main() {
+    println!("cargo:rerun-if-changed=build.rs");
+    build_platform_cfg::emit();
+}
